Raise an exception when we can't locate the virtualenv

+++ yola/deploy/hooks/python.py
@@ -189,7 +189,9 @@ class PythonApp(DeployHook):
 
         if not os.path.exists(ve_working):
             os.makedirs(ve_working)
-        download_ve(self.app, ve_hash, self.artifacts_factory, tarball)
+        if not download_ve(self.app, ve_hash, self.artifacts_factory,
+                           tarball):
+            raise Exception("Could not locate virtualenv %s" % ve_hash)
         unpack_ve(tarball, ve_working)
         if os.path.exists(ve_dir):
             shutil.rmtree(ve_dir)
